Refactor UI and add batch features
